Base response models moved to base package
