+++ src/main/java/com/cyberfreak/services/api/controller/ApplicationController.java
@@ -4,10 +4,10 @@ import com.cyberfreak.services.api.constants.ApiEndpoints;
 import com.cyberfreak.services.api.constants.ApiPaths;
 import com.cyberfreak.services.api.request.application.CreateApplicationRequest;
 import com.cyberfreak.services.api.request.application.UpdateApplicationRequest;
-import com.cyberfreak.services.api.response.BaseResponse;
-import com.cyberfreak.services.api.response.ListResultResponse;
-import com.cyberfreak.services.api.response.SaveEntityResponse;
-import com.cyberfreak.services.api.response.SingleResultResponse;
+import com.cyberfreak.services.api.response.base.BaseResponse;
+import com.cyberfreak.services.api.response.base.ListResultResponse;
+import com.cyberfreak.services.api.response.base.SaveEntityResponse;
+import com.cyberfreak.services.api.response.base.SingleResultResponse;
 import com.cyberfreak.services.dto.ApplicationDto;
 import com.cyberfreak.services.service.ApplicationService;
 import lombok.RequiredArgsConstructor;
