+++ src/main/java/com/cyberfreak/services/api/controller/PageContentController.java
@@ -6,9 +6,9 @@ import com.cyberfreak.services.api.request.pagecontent.AddContentItemsRequest;
 import com.cyberfreak.services.api.request.pagecontent.CreatePageContentRequest;
 import com.cyberfreak.services.api.request.pagecontent.CreatePageContentWithExistingItemsRequest;
 import com.cyberfreak.services.api.request.pagecontent.CreatePageContentWithItemsRequest;
-import com.cyberfreak.services.api.response.ListResultResponse;
-import com.cyberfreak.services.api.response.SaveEntityResponse;
-import com.cyberfreak.services.api.response.SingleResultResponse;
+import com.cyberfreak.services.api.response.base.ListResultResponse;
+import com.cyberfreak.services.api.response.base.SaveEntityResponse;
+import com.cyberfreak.services.api.response.base.SingleResultResponse;
 import com.cyberfreak.services.dto.PageContentDto;
 import com.cyberfreak.services.service.PageContentService;
 import lombok.RequiredArgsConstructor;
