+++ src/main/java/com/cyberfreak/services/api/response/base/ListResultResponse.java
@@ -1,4 +1,4 @@
-package com.cyberfreak.services.api.response;
+package com.cyberfreak.services.api.response.base;
 
 import com.cyberfreak.services.dto.base.BaseDto;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
@@ -10,15 +10,12 @@ import java.util.List;
 @Getter
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class PagedListResultResponse<T extends BaseDto> extends BaseResponse {
+public class ListResultResponse<T extends BaseDto> extends BaseResponse {
 
     private List<T> resultList;
 
-    private PageData pageData;
-
-    public PagedListResultResponse(List<T> resultList, PageData pageData) {
+    public ListResultResponse(List<T> resultList) {
         super(true);
         this.resultList = resultList;
-        this.pageData = pageData;
     }
 }
