+++ src/main/java/com/cyberfreak/services/api/response/base/PagedListResultResponse.java
@@ -1,4 +1,4 @@
-package com.cyberfreak.services.api.response;
+package com.cyberfreak.services.api.response.base;
 
 import com.cyberfreak.services.dto.base.BaseDto;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
@@ -10,12 +10,15 @@ import java.util.List;
 @Getter
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ListResultResponse<T extends BaseDto> extends BaseResponse {
+public class PagedListResultResponse<T extends BaseDto> extends BaseResponse {
 
     private List<T> resultList;
 
-    public ListResultResponse(List<T> resultList) {
+    private PageData pageData;
+
+    public PagedListResultResponse(List<T> resultList, PageData pageData) {
         super(true);
         this.resultList = resultList;
+        this.pageData = pageData;
     }
 }
